+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 2.9.14
 - _makeMultipleInput(): Fix div ID in StudentsUsersInfo.fnc.php
 - Fix Save Select multiple from options field in User.php
 - Admin User Profile restriction in Profiles.php, Exceptions.php, User.php & General_Info.inc.php (sponsored by Aptiris)
+- Update database for Admin User Profile restriction in rosariosis.sql & Update.fnc.php
 
 Changes in 2.9.13
 -----------------